Web/ExcessCookie_V2/includes/flaskapp/requirements.txt
Flask
Flask-PyMongo
Werkzeug
python-decouple
waitress
gunicorn
bcrypt
pyjwt
pymongo
Flask-reCaptcha